requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==24.1.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
bidict==0.23.1
boto3==1.35.85
botocore==1.35.85
CacheControl==0.12.14
cachetools==4.2.4
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
cloudpickle==1.6.0
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==43.0.3
cycler==0.12.1
dask==2022.2.0
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.7.0
flatbuffers==2.0.7
fsspec==2024.12.0
future==1.0.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httptools==0.6.4
humanfriendly==10.0
idna==3.10
importlib-metadata==8.5.0
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==4.4.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
locket==1.0.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.1.0
multidict==5.2.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==20.9
pamqp==2.3.0
partd==1.4.2
pillow==11.0.0
prompt-toolkit==2.0.10
propcache==0.2.1
protobuf==3.19.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydot==1.4.2
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
pytelegrambotapi==4.25.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.11.1
python-socketio==5.12.0
pytz==2021.3
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.2.0
rasa-sdk==3.2.3
redis==3.5.3
regex==2022.4.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.25.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
sanic==21.12.1
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
sentry-sdk==1.3.1
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
SQLAlchemy==1.4.54
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
toolz==1.0.0
tqdm==4.67.1
twilio==6.50.1
typeguard==4.4.1
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==5.10.0
urllib3==1.26.6
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==14.1
werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
